Make sure to install profile package before build.
